feat(wfsk): W-FSK — 11 paralelních 16-FSK skupin, 2,75 kbit/s
Přidává wideband variantu Q-FSK: skupiny G5–G11 (7500–15000 Hz) nad
původní G1–G4, celkem 11 paralelních 16-FSK skupin v pásmu ~1–16 kHz.
44 bitů/symbol → 2,75 kbit/s @ 62,5 Bd (2,75× Q-FSK). Horní pásma podle
vysokofrekvenční sondáže kanálu na macu (čistý průchod až do ~16 kHz).

Změny:
- qfsk.cpp zparametrizováno na seznam bází (NfskModulator/NfskDemodulator),
  sdílený kód pro Q-FSK (kQfskBase, 4) i W-FSK (kWfskBase, 11).
- Symbolová cesta rozšířena na uint64_t (44 b/symbol > 32): BitBuffer::
  symbol()/pushBits(), IDemodulator::demodSymbol() (všech 5 schémat),
  FrameReceiver::takeSymbol() + onSymbol hook. Ostatní schémata beze změny.
- Registrace "W-FSK" (registry.cpp, schemes.hpp); default baud 62,5 v CLI
  (applyCommonConfig) i GUI (parseGuiArgs, combo v panels.cpp).
- Testy: 4 W-FSK case (44 b/symbol, propustnost 2750, max-payload round-trip
  přes uint64 cestu, šum+drift); chain testy pokrývají W-FSK přes registr.
  50/50 case, 1231 assertions, PASS.
- README: tabulka modulací + popis W-FSK + ověřené výsledky.

Ověřeno akusticky self-loopbackem na macu: BER 0/3072, SNR ~25 dB při
2,75 kbit/s (2× nezávisle). Horní pásma zatím jen na macu — Fedora→Mac
klesá nad ~5,5 kHz, dvoustrojové měření W-FSK čeká. Výchozí robustní
volbou zůstává Q-FSK.

+++ src/core/bits.hpp
@@ -16,7 +16,7 @@ public:
     void pushBit(bool b) { bits_.push_back(b ? 1 : 0); }
 
     // n nejnižších bitů hodnoty v, LSB první
-    void pushBits(uint32_t v, int n) {
+    void pushBits(uint64_t v, int n) {
         for (int i = 0; i < n; ++i) pushBit((v >> i) & 1u);
     }
 
@@ -29,10 +29,11 @@ public:
     size_t size() const { return bits_.size(); }
     bool   empty() const { return bits_.empty(); }
 
-    // i-tá skupina n bitů jako číslo (pro M-ární modulace), LSB první
-    uint32_t symbol(size_t i, int n) const {
-        uint32_t v = 0;
-        for (int k = 0; k < n; ++k) v |= uint32_t(bits_[i * n + k]) << k;
+    // i-tá skupina n bitů jako číslo (pro M-ární modulace), LSB první.
+    // Návratový typ uint64_t → schémata s >32 b/symbol (W-FSK: 44 b).
+    uint64_t symbol(size_t i, int n) const {
+        uint64_t v = 0;
+        for (int k = 0; k < n; ++k) v |= uint64_t(bits_[i * n + k]) << k;
         return v;
     }
